Updated to revision 1.5 and Second Edition edits

+++ uMB/VS1053.c
@@ -13,7 +13,7 @@ WORD  writeMP3( BYTE b)
     return MP3_SPIBUF;          // return buffer content
 }
 
-void writeMP3Register( BYTE reg, unsigned w)
+void writeMP3Register( BYTE reg, WORD w)
 {
     MP3_DCS_Disable();              // disable data bus
     MP3_CS_Enable();                // enable command bus
